move array manipulation data to util function
